requirements.txt
python-telegram-bot==13.5
SQLAlchemy~=2.0.30
bs4~=0.0.2
requests~=2.32.3
pytube~=15.0.0
Pillow>=8.1.1
telethon~=1.36.0
spotipy~=2.24.0
beautifulsoup4~=4.12.3
instaloader~=4.11
jdatetime~=5.0.0
python-dateutil~=2.9.0.post0